Port to native cytoscape integration
