+++ timesketch/frontend-ng/package.json
@@ -9,20 +9,14 @@
     "test:unit": "vue-cli-service test:unit"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.7.2",
-    "@fortawesome/fontawesome-svg-core": "^1.2.31",
-    "@fortawesome/free-solid-svg-icons": "^5.7.2",
-    "@fortawesome/vue-fontawesome": "^2.0.0",
     "@mdi/font": "^7.1.96",
     "apexcharts": "^3.27.1",
     "axios": "^0.21.1",
-    "buefy": "^0.9.4",
-    "bulma": "^0.9.1",
-    "cytoscape-dagre": "^2.3.1",
+    "cytoscape": "^3.24.0",
+    "cytoscape-dagre": "^2.5.0",
     "cytoscape-spread": "^3.0.0",
     "dayjs": "^1.10.7",
     "dompurify": "^3.0.1",
-    "eslint-config-prettier": "^8.3.0",
     "lodash": "^4.17.11",
     "marked": ">=2.0.0 <5.0.0",
     "moment": "^2.24.0",
@@ -32,15 +26,9 @@
     "vega-lite": "3.4.0",
     "vue": "^2.6.12",
     "vue-apexcharts": "^1.6.1",
-    "vue-clipboard2": "^0.3.1",
-    "vue-color": "^2.7.0",
-    "vue-cytoscape": "^1.0.8",
     "vue-dragscroll": "^3.0.0",
     "vue-moment": "^4.0.0",
-    "vue-multiselect": "^2.1.6",
     "vue-router": "^3.0.2",
-    "vue-scrollto": "^2.17.1",
-    "vue-text-highlight": "^2.0.10",
     "vuetify": "^2.5.14",
     "vuex": "^3.1.0"
   },
